requirements.txt
asgiref==3.7.2
Django==4.2.3
django-cors-headers==4.1.0
djangorestframework==3.14.0
djongo==1.2.31
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.4
